+++ src/com/osudroid/ui/v2/menu/main/Logo.kt
@@ -1,4 +1,4 @@
-package com.acivev.ui.menu.main
+package com.osudroid.ui.v2.menu.main
 
 import android.util.Log
 import com.reco1l.andengine.Anchor
