+++ src/com/osudroid/ui/v2/menu/main/MusicOption.kt
@@ -1,4 +1,4 @@
-package com.acivev.ui.menu.main
+package com.osudroid.ui.v2.menu.main
 
 enum class MusicOption { PREV, PLAY, PAUSE, STOP, NEXT, SYNC }
